Give aria refs one home
The ref helpers had grown across three files: clickRef's selector, staleness
check and element description sat among the general tools, the ai-mode
snapshot lived with the eidx annotator, and two unused statics hung off
WebElement. They are all one concept, so they are now one module.

web-annotate keeps what is actually its own — which roles it annotates — and
reads the refs out of a snapshot through the shared parser. No behaviour
changes; describeRef passes its tag-to-role map into the page rather than
closing over it, since the map is now module scope.
